Stop using UrlVariables since it is un-ordered.
UrlVariables will output the params in random order. 99% of the time this
doesn't matter but occasionally it does. This removes the use of UrlVariables.
It is now just a simple list of params and when doing a UrlRequest I encode it
myself.

This has the side benefit of getting rid of the recursive array logic. Params
are not just flat and if someone want to encode an array they can do so without
us caring since we don't have to worry about keys overwritting each other.