Better solution for filter by file extension
by adding the browsers that currently support it to the capabilities
tester of the html5, html4 runtimes.

This is part of the W3C spec, as documented here:
http://www.w3.org/TR/2012/WD-html5-20121025/states-of-the-type-attribute
.html#attr-input-accept

At the moment only some browsers support filtering by extension
(Safari, Chrome, IE10, IE11). But hopefully support for other browsers
should be in the pipeline.

Please note: If/when support for other browsers is implemented,
capabilities tester of both runtimes will need to be updated.